BUG/MINOR: config: disable http-reuse on TCP proxies
Louis Chanouha reported an inappropriate warning when http-reuse is
present in a defaults section while a TCP proxy accidently inherits
it and finds a conflict with other options like the use of the PROXY
protocol. To fix this patch removes the http-reuse option for TCP
proxies.

This fix needs to be backported to 1.8, 1.7 and possibly 1.6.

(cherry picked from commit 46deab6e64bfda7211b7c3199ad01f136141c86f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>