BUG/MEDIUM: servers: Add srv_addr default placeholder to the state file
When creating a state file using "show servers state" an empty field is
created in the srv_addr column if the server is from the socket family
AF_UNIX.  This leads to a warning on start up when using
"load-server-state-from-file". This patch defaults srv_addr to "-" if
the socket family is not covered.

This patch should be backported to 1.8.

(cherry picked from commit 9215ffa6b294374e99e5ab027bea9b34642d7277)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>