BUG/BUILD: fd: fix typo causing a warning when threads are disabled
A warning is issued since commit 954db1d ("BUG/MEDIUM: pollers: Use a
global list for fd shared between threads.") when threads are disabled
due to a typo. It has no effect beyond this warning.

There is no mainline commit ID because this bug is specific to the
backport above.