BUG/MEDIUM: lua/socket: wrong scheduling for sockets
The appctx pointer is given from any variable which are wrong.
This implies the wakeup of wrong applet, and the socket are no
longer responsive.

This behavior is hidden by another inherited error which is
fixed in the next patch.

This patch remove all wrong appctx affectations.

This patch must be backported in 1.6, 1.7 and 1.8

(cherry picked from commit af4bd0867a6ef4ca219bf97fee5205fa928be4ab)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>