BUG/MEDIUM: lua/socket: Length required read doesn't work
The limit of data read works only if all the data is in the
input buffer. Otherwise (if the data arrive in chunks), the
total amount of data is not taken in acount.

Only the current read data are compared to the expected amout
of data.

This patch must be backported from 1.9 to 1.6

(cherry picked from commit 8c126c7235ced98986a57175a62503d5166ffe19)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>