BUG/MEDIUM: lua/socket: Notification error
Each time the send function yields, a notification must be registered.
Without this notification, the task is never wakeup when data arrives.

Today, the notification is registered only if the buffer is not available.
Other cases like the buffer is too small for all data are not processed.

This patch must be backported in 1.6, 1.7 and 1.8

(cherry picked from commit ba42fcd064c3f371610126dee6f25c2d7f1e4040)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>