BUG/MEDIUM: fd: Only check update_mask against all_threads_mask.
To decide if we have to insert a fd into the update_list, we check its
update_mask. We set the update_mask to fdtab[fd].thread_mask when adding it
or updating, and when each thread will remove its bit once it has been taken
care of.
However, even if all threads did their job, update_mask may be non-zero,
because for fds that want all threads, the mask will be 0xffffffffffffffff,
so except when we run with 64 threads, there will be bits left over.

So, to decide if nobody had any update for that fd, we have to check
(update_mask & all_threads_mask) == 0, instead of update_mask == 0. Not
doing so means fd events can be missed, or fd not removed from the poller,
making poller waking with no need, and using 100% of the CPU.

This should only be applied on 1.8.