BUG/MINOR: contrib/spoa_example: Don't reset the status code during disconnect
When the connection is closed by HAProxy, the status code provided in the
DISCONNECT frame is lost. By retransmitting it in the agent's reply, we are sure
to have it in the SPOE logs.

This patch may be backported in 1.8.

(cherry picked from commit b47e4385933b645cf6cac846599feb8a823822d1)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>