BUG/MINOR: contrib/mod_defender: Don't reset the status code during disconnect
When the connection is closed by HAProxy, the status code provided in the
DISCONNECT frame is lost. By retransmitting it in the agent's reply, we are sure
to have it in the SPOE logs.

This patch may be backported in 1.8.

(cherry picked from commit 72de8eb7aa4c2e890ed1a5c89d40b8b041c71b0c)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>