BUG/MINOR: contrib/modsecurity: Don't reset the status code during disconnect
When the connection is closed by HAProxy, the status code provided in the
DISCONNECT frame is lost. By retransmitting it in the agent's reply, we are sure
to have it in the SPOE logs.

This patch may be backported in 1.8.

(cherry picked from commit 58b8e492a311b5dfec533734c1e22ea76d352aed)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>