BUG/MAJOR: map: fix a segfault when using http-request set-map
The bug happens with an existing entry, when you try to overwrite the
value with wrong data, for example, a string when the type is INT.

The code path was not secure and tried to set *err and *merr while
err = merr = NULL when performing an http action.

Must be backported in 1.6, 1.7, 1.8.

(cherry picked from commit 579fb25b6215d64a2e064550c1d186547414cc68)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>