BUG/MINOR: servers: Don't make "server" in a frontend fatal.
When parsing the configuration, if "server", "default-server" or
"server-template" are found in a frontend, we first warn that it will be
ignored, only to be considered a fatal error later. Be true to our word, and
just ignore it.

This should be backported to 1.8 and 1.7.

(cherry picked from commit 306e653331f107902f6211f03e09779cfe0171d0)
Signed-off-by: Willy Tarreau <w@1wt.eu>